Chapter09/requirements.txt
numpy
Pillow
tensorflow
tensorflow-hub
requests